+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Other changes made
 - Removed the '/hideout' from the list of default commands (It can now be assigned a hotkey in-game instead)
+- Added support for 'Enshrouded' and 'Chart' item categories
+- Added 'Ducat' to the list of currencies to obtain from poe ninja
+- Added support for 'Sanctum Relics' and 'Enshrouded' exchange rates (obtained from poe ninja)
 
 ### Changes made when merging master v0.8.42 into this branch (which originated from v0.8.39)
 - Removed the polyfill workaround for poe login page